slack-connector/pyproject.toml
[project]
name = "slack-connector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.2",
    "python-dotenv>=1.1.1",
    "slack-sdk>=3.36.0",
]
